refactor: move existing shortcodes into own directory
